Clean up Chip RAM BUFEN signal. 23-JUN-2024
The _BUFEN signal was negated too soon. This was caused by how the internal TS signal was negated. The TS signal reset was changed in U409 to allow _BUFEN to stay asserted for a more appropriate length of time. This change also has the positive effect of allowing the logic to respond sooner to new CPU data transfer cycles.