U409: Cleaned up ROM cycle.
When changing addresses, the 27C4096 ROM drives all the data bits high. This makes noisy data transitions between addresses when output enable is held asserted for back-to-back ROM cycles. This was causing instability in the Rev 5.0 hardware. The new ROM cycle disables output from the ROM between back-to-back ROM cycles.